Moved graphql_menu to graphql_core.
