+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Menu/MenuLinkLabel.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace Drupal\graphql_menu\Plugin\GraphQL\Fields;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Menu;
 
 use Drupal\Core\Menu\MenuLinkTreeElement;
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
 /**
- * Menu link descriptions.
+ * Retrieve a menu links label.
  *
  * @GraphQLField(
- *   id = "menu_link_description",
+ *   id = "menu_link_label",
  *   secure = true,
- *   name = "description",
+ *   name = "label",
  *   type = "String",
  *   types = {"MenuLink"}
  * )
  */
-class MenuLinkDescription extends FieldPluginBase {
+class MenuLinkLabel extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   public function resolveValues($value, array $args, ResolveInfo $info) {
     if ($value instanceof MenuLinkTreeElement) {
-      yield $value->link->getDescription();
+      yield $value->link->getTitle();
     }
   }
