+++ modules/graphql_core/tests/queries/menu.gql
