+++ modules/graphql_menu/graphql_menu.info.yml
@@ -1,6 +1,7 @@
 name: GraphQL Menu
 type: module
 description: 'Query Drupal menu trees with GraphQL.'
+hidden: true
 package: GraphQL
 core: 8.x
 dependencies:
